refactored firstcollisionmnsolver.h
