+++ include/solvers/firstcollisionmnsolver.h
